Fix format issue
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@1692 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/Common/host.c
@@ -1069,7 +1069,8 @@ save_recent(const char *hn)
 		    ctime(&t), build);
 	    for (h = hosts; h != NULL; h = h->next) {
 		if (h->entry_type == RECENT) {
-		    (void) fprintf(lcf, "%lu %s\n", h->connect_time, h->name);
+		    (void) fprintf(lcf, "%lu %s\n",
+			    (unsigned long)h->connect_time, h->name);
 		}
 	    }
 	    fclose(lcf);
